Change: Refactor 'logs' module in 'local' sub-package for more clear struct.
